Move ModScreens from client.init -> init.client
