+++ src/main/java/xyz/iwolfking/woldsvaults/events/client/ClientSetupEvents.java
@@ -7,12 +7,13 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import xyz.iwolfking.woldsvaults.init.client.ModEntityRenderers;
 import xyz.iwolfking.woldsvaults.init.client.ModKeybinds;
+import xyz.iwolfking.woldsvaults.init.client.ModScreens;
 
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD, value = {Dist.CLIENT})
 public class ClientSetupEvents {
     @SubscribeEvent(priority = EventPriority.LOW)
     public static void setupClient(FMLClientSetupEvent event) {
-        xyz.iwolfking.woldsvaults.client.init.ModScreens.register();
+        ModScreens.register();
         ModEntityRenderers.register(event);
         ModKeybinds.registerKeyBinds();
     }
